# analytics-engine/openapi/analytics-engine-api.yaml
openapi: 3.0.3
info:
  title: Analytics Engine API
  description: |
    Production-ready Analytics Engine API for AI video analytics, human tracking,
    vehicle analytics, industrial safety, and real-time detection processing.
    
    **Core Features:**
    - Real-time frame processing with ONNX inference
    - 14 analytics modules (human, vehicle, industrial, retail, banking, etc.)
    - Person/vehicle tracking with Re-ID
    - Heat map generation and crowd analytics
    - Industrial safety monitoring
    - SNMP device integration
    - Camera health monitoring
    
    **Performance:**
    - <100ms person detection latency
    - 1-2 FPS frame processing
    - 30+ concurrent camera streams (CPU)
    - 50-80+ streams (GPU)
    
    **Authentication:**
    All endpoints require `x-analytics-source-key` or `x-analytics-engine-key` header.
  version: 2.0.0
  contact:
    name: Analytics Engine Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8092
    description: Development server
  - url: https://analytics.sentinelgrid.com
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/analytics/cameras/{cameraId}/status:
    get:
      tags: [Health]
      summary: Get camera AI status
      description: Get per-camera AI analytics status with inference metrics
      operationId: getCameraStatus
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Camera status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraStatus'
  /internal/detections:
    post:
      tags: [Detection]
      summary: Submit detection event
      description: Submit a single detection event to control plane
      operationId: submitDetection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetectionEvent'
      responses:
        '202':
          description: Detection accepted
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          description: Control plane unavailable
  /internal/detections/batch:
    post:
      tags: [Detection]
      summary: Submit batch detection events
      description: Submit up to 100 detection events in a single request
      operationId: submitDetectionBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/DetectionEvent'
              minItems: 1
              maxItems: 100
      responses:
        '202':
          description: Batch accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: integer
                  failed:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        sourceEventId:
                          type: string
                        status:
                          type: string
                          enum: [accepted, failed]
  /internal/frames:
    post:
      tags: [Detection]
      summary: Process frame
      description: |
        Submit RGB24 frame for local ONNX inference or provide pre-computed detections.
        Omit `detections` to run local inference. Provide `detections: []` to indicate
        upstream worker found no objects.
      operationId: processFrame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FrameInput'
      responses:
        '202':
          description: Frame processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  cameraId:
                    type: string
                  inferenceMode:
                    type: string
                    enum: [local-onnx, normalized-observation]
                  detectionsReceived:
                    type: integer
                  eventsGenerated:
                    type: integer
                  accepted:
                    type: integer
                  failed:
                    type: integer
  /v1/detections/persons/tracks:
    get:
      tags: [Human Analytics]
      summary: Get active person tracks
      description: Get currently tracked persons across all cameras
      operationId: getPersonTracks
      parameters:
        - name: cameraId
          in: query
          schema:
            type: string
        - name: minDwellSeconds
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Active tracks
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  tracks:
                    type: array
                    items:
                      $ref: '#/components/schemas/PersonTrack'
...
  /v1/detections/vehicles/tracks:
    get:
      tags: [Vehicle Analytics]
      summary: Get active vehicle tracks
      description: Get currently tracked vehicles with speed and direction
      operationId: getVehicleTracks
      parameters:
        - name: cameraId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Active tracks
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  tracks:
                    type: array
                    items:
                      $ref: '#/components/schemas/VehicleTrack'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-analytics-source-key
      description: Analytics engine API key
  schemas:
...
    CameraStatus:
      type: object
      properties:
        cameraId:
          type: string
        stream:
          type: string
          enum: [connected, disconnected, unknown]
        recording:
          type: boolean
        aiEngine:
          type: string
          enum: [AI_OPERATIONAL, AI_DEGRADED, AI_UNAVAILABLE]
        model:
          type: string
          example: "YOLOX Tiny 0.1.1"
        inferenceMode:
          type: string
          enum: [local-onnx, normalized-observation]
        inference:
          type: object
          properties:
            fps:
              type: number
            latencyMs:
              type: number
            gpu:
              type: object
        detection:
          type: object
          properties:
            persons:
              type: integer
            vehicles:
              type: integer
            lastDetectionAt:
              type: string
              format: date-time
    DetectionEvent:
      type: object
      required: [tenantId, cameraId, detectionType, confidence, modelVersion]
      properties:
        tenantId:
          type: string
        cameraId:
          type: string
        sourceEventId:
          type: string
        detectionType:
          type: string
        occurredAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        confidence:
          type: number
          minimum: 0
          maximum: 1
        durationSeconds:
          type: number
        modelVersion:
          type: string
        objects:
          type: array
          items:
            $ref: '#/components/schemas/DetectedObject'
        snapshotReference:
          type: string
        clipReference:
          type: string
        metadata:
          type: object
    DetectedObject:
      type: object
      required: [label, confidence]
      properties:
        label:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        trackId:
          type: string
        boundingBox:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            width:
              type: number
            height:
              type: number
    FrameInput:
      type: object
      required: [tenantId, cameraId, width, height]
      properties:
        tenantId:
          type: string
        cameraId:
          type: string
        capturedAt:
          type: string
          format: date-time
        width:
          type: integer
        height:
          type: integer
        imageBase64:
          type: string
          format: byte
          description: RGB24 image data (width * height * 3 bytes)
        detections:
          type: array
          items:
            $ref: '#/components/schemas/DetectedObject'
          description: Pre-computed detections (omit to run local inference)
        rules:
          type: array
          items:
            type: object
        metadata:
          type: object
    PersonTrack:
      type: object
      properties:
        trackId:
          type: string
        firstSeen:
          type: string
          format: date-time
        lastSeen:
          type: string
          format: date-time
        dwellTimeSeconds:
          type: number
        isStationary:
          type: boolean
        positionHistory:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number
              timestamp:
                type: string
                format: date-time
    VehicleTrack:
      type: object
      properties:
        trackId:
          type: string
        vehicleType:
          type: string
        speed:
          type: number
        direction:
          type: string
        licensePlate:
          type: string
        firstSeen:
          type: string
          format: date-time
        lastSeen:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: object
...